Removes expit normalization from OnnxDetectionPredictor
While a custom post-processor can reverse the normalization back to the
original values via logit, it would make more sense to leave
normalization to the post-processor. This way the post-processor gets
the raw values from the model, which is easier to reason with, instead
of being forced to run `logit` on everything to get raw values...

Additionally, with this change there should be fewer expit calls, as
we only need it for the box prediction scores now.

THIS IS A BACKWARDS COMPATIBILITY BREAKAGE! If a user made a custom
detection post-processor, it will start getting different values in the
buffer than before!

Autoported commit.
Original commit hash: [7244cf9]
Manual files:
pdfocr-onnxtr/src/main/java/com/itextpdf/pdfocr/onnxtr/detection/OnnxDetectionPostProcessor.java